Add Windows manifest to enable long paths (#649)
While we could have used a target-specific build dependency, it confuses Cargo.lock. So for now we force all OSes to download the winres crate (but the stuff in the crate should probably be no-ops on non-Windows). Resolves #646 to the extent where possible. Users still need to opt-in to the new behaviour by setting a registry key (https://docs.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation).